SEARCH-1358: redirect from old aleph_id to new alma id (#25)
If we get a 9-digit id on a mirlyn single-record query, change the id_field for both the record query and the holdings query to aleph_id.

There's an apparently extraneous `id_field ||= 'id'` line in BibRecord.fetch because if that error ever comes up (someone passes in an id_field that's nil) it will be super annoying to diagnose. As little behavior change as possible here.